requirements.txt
pyyaml
fire
jinja2
attrs